Module Xml_printer

type xml = Xml_datatype.xml

type t

type target =

| TChannel of Stdlib.out_channel

| TBuffer of Stdlib.Buffer.t

val make : target -> t

val print : t -> xml -> unit

Print the xml data structure to a source into a compact xml string (without any user-readable formatting ).

val to_string : xml -> string

Print the xml data structure into a compact xml string (without any user-readable formatting ).

val to_string_fmt : xml -> string

Print the xml data structure into an user-readable string with tabs and lines break between different nodes.

val pcdata_to_string : string -> string

Print PCDATA as a string by escaping XML entities.